np-41723: moved new version model to package with version name
