+++ cristin-organization/src/main/java/no/unit/nva/cristin/organization/CristinOrgApiClientVersion2.java
@@ -12,8 +12,8 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import no.unit.nva.cristin.common.client.IQueryApiClient;
 import no.unit.nva.cristin.model.SearchResponse;
-import no.unit.nva.cristin.organization.dto.OrganizationFromUnitMapper;
-import no.unit.nva.cristin.organization.dto.UnitDto;
+import no.unit.nva.cristin.organization.dto.version20230510.mapper.OrganizationFromUnitMapper;
+import no.unit.nva.cristin.organization.dto.version20230510.UnitDto;
 import no.unit.nva.model.Organization;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
 import nva.commons.apigateway.exceptions.BadGatewayException;
