+++ cristin-organization/src/main/java/no/unit/nva/cristin/organization/dto/version20230510/mapper/OrganizationFromUnitMapper.java
@@ -1,4 +1,4 @@
-package no.unit.nva.cristin.organization.dto;
+package no.unit.nva.cristin.organization.dto.version20230510.mapper;
 
 import static java.util.Objects.nonNull;
 import static no.unit.nva.cristin.model.Constants.ORGANIZATION_PATH;
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import no.unit.nva.cristin.organization.dto.version20230510.UnitDto;
 import no.unit.nva.model.Organization;
 
 public class OrganizationFromUnitMapper implements Function<UnitDto, Organization> {
